# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ciudadano AI
  version: 2.0.0
  description: 'Prototipo academico. Profile: datos propios. Citizen: seguimiento
    OTP. Admin: permisos. Las rutas /internal no se publican.'
servers:
- url: http://localhost:8000
paths:
...
  /api/v1/models/current:
    get:
      summary: Backend explicito y metricas disponibles
      responses:
        '200':
          description: Operacion completada
        '400':
          description: Entrada invalida
        '401':
          description: Token ausente o invalido
        '403':
          description: Permiso, titularidad o consentimiento insuficiente
        '429':
          description: Limite de tasa o intentos
        '503':
          description: Dependencia no disponible
...
  /api/v1/classify:
    post:
      summary: Clasificar con backend declarado
      responses:
        '200':
          description: Operacion completada
        '400':
          description: Entrada invalida
        '401':
          description: Token ausente o invalido
        '403':
          description: Permiso, titularidad o consentimiento insuficiente
        '429':
          description: Limite de tasa o intentos
        '503':
          description: Dependencia no disponible
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  minLength: 1
                  maxLength: 1500
              required:
              - text
  /api/v1/extract-entities:
    post:
      summary: Extraer entidades
      responses:
        '200':
          description: Operacion completada
        '400':
          description: Entrada invalida
        '401':
          description: Token ausente o invalido
        '403':
          description: Permiso, titularidad o consentimiento insuficiente
        '429':
          description: Limite de tasa o intentos
        '503':
          description: Dependencia no disponible
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  minLength: 1
                  maxLength: 1500
              required:
              - text
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...